+++ index.js
@@ -12,3 +12,4 @@ export * as M from './src/maybe.js';
 export * as E from './src/either.js';
 export * as T from './src/transformations.js';
 export * as C from './src/cancelable.js';
+export * as L from './src/list.js';